backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
supabase==2.9.1
google-genai>=0.3.0
pinecone-client==5.0.1
pydantic==2.9.2
httpx==0.27.2
python-multipart==0.0.9
Pillow==10.4.0
exponent-server-sdk==2.2.0
posthog==3.4.0
apscheduler==3.10.4
redis==5.0.1
pytest==7.4.3
pytest-asyncio==0.21.1

